+++ src/Entity/Tile.php
@@ -106,7 +106,8 @@ class Tile
 
     public function getImage(): ?string
     {
-        return $this->image;
+        // HACK: add path here prefix to ensure image paths work both in API and easy admin.
+        return '/tiles/'.$this->image;
     }
 
     public function setImage(string $image): self